requirements.txt
attrs==23.1.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
-e git+ssh://git@github.com/Kromag1995/scanHouse.git@f08476e593a66f81b57c711ba709966153d2192a#egg=dashV&subdirectory=dashV
fake-useragent==1.1.3
Faker==18.4.0
filelock==3.12.0
Flask==2.2.3
greenlet==2.0.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parsel==1.8.1
plotly==5.14.1
-e git+ssh://git@github.com/Kromag1995/scanHouse.git@f08476e593a66f81b57c711ba709966153d2192a#egg=props_dashboard&subdirectory=dashV
Protego==0.2.1
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
scrapy-fake-useragent==1.4.4
scrapy-splash==0.9.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.10
tenacity==8.2.2
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
